// src/types/supabase.ts
export interface definitions {
// ...
  owner_infomation: {
    /**
     * Format: uuid
     * @description Note:
     * This is a Primary Key.<pk/>
     */
    id: string;
    /**
     * Format: timestamp with time zone
     * @default now()
     */
    updatedAt?: string;
    /**
     * Format: timestamp with time zone
     * @default now()
     */
    createdAt?: string;
    /** Format: text */
    name?: string;
    /** Format: text */
    apiChannelId?: string;
    /** Format: text */
    liffChannelSecret?: string;
    /** Format: text */
    webhookUrl?: string;
    /** Format: text */
    liffId?: string;
    /** Format: text */
    liffChannelId?: string;
    /** Format: text */
    apiChannelSecret?: string;
    /** Format: text */
    apiChannelAccessToken?: string;
    /** Format: text */
    basicId?: string;
    /** Format: text */
    pictureUrl?: string;
    /** Format: text */
    liffAccessToken?: string;
    /**
     * Format: boolean
     * @default true
     */
    firstLogin: boolean;
    /** Format: boolean */
    agreeTerms?: boolean;
  };
// ...
}